typescript(vx-demo): rewrite Dots example in TypeScript
