+++ packages/vx-demo/src/components/gallery.js
@@ -8,7 +8,7 @@ import Footer from './Footer.tsx';
 
 import Lines from './tiles/lines';
 import Bars from './tiles/bars';
-import Dots from './tiles/dots';
+import Dots from './tiles/Dots.tsx';
 import Patterns from './tiles/Patterns.tsx';
 import Gradients from './tiles/Gradients.tsx';
 import Area from './tiles/area';
